Create fibonacci
Here's a C program that generates the Fibonacci series up to a given number `n`:

```c
#include <stdio.h>

void generateFibonacci(int n) {
    int first = 0, second = 1, next;

    printf("Fibonacci Series up to %d:\n", n);
    printf("%d, %d, ", first, second);

    next = first + second;
    while (next <= n) {
        printf("%d, ", next);
        first = second;
        second = next;
        next = first + second;
    }

    printf("\n");
}

int main() {
    int n;

    printf("Enter the number up to which Fibonacci series should be generated: ");
    scanf("%d", &n);

    generateFibonacci(n);

    return 0;
}
```

This program defines a function `generateFibonacci` that takes a number `n` as input and generates the Fibonacci series up to that number. It uses a while loop to iterate and generate the Fibonacci numbers until the next number exceeds `n`. Finally, the `main` function prompts the user to enter the value of `n`, calls the `generateFibonacci` function, and prints the Fibonacci series up to that number.